% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quotes.R
\name{add_quote}
\alias{add_quote}
\title{Interval Closure Classification}
\usage{
add_quote(x, qmark = "'")
}
\arguments{
\item{x}{any character object to be be surrounded with quotation marks. If
x is not a character object and can be coerced to a character object, it
will, but with a warning message.}

\item{qmark}{Any character to 'surround' x. Defaults to single quotation
marks (not fancy).}
}
\value{
a character vector in single quotes (default behavior). Changing
qmark will result in different character(s) surrouding 'x'.
}
\description{
An alternative to sQuote, ideally for passing quoted values to
another R function. When you want fancy quotes (for visual appeal),
use sQuote.
}
\note{
sQuote is very similar but to get the same behavior options must
be changed. (i.e. options(useFancyQuotes = FALSE))
}
\examples{
add_quote("10/1/2012")
add_quote(1)

}

