% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmaps.R
\name{brmaps}
\alias{brmaps}
\title{Function to read Brazilian municipalities map for a year}
\usage{
brmaps(y, rfolder = "~/maps/br/")
}
\arguments{
\item{y}{integer for year}

\item{rfolder}{the root folder}
}
\value{
sf object with first and second column
as geocode and name
}
\description{
Function to read Brazilian municipalities map for a year
}
