% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs2f.R
\name{bs2f}
\alias{bs2f}
\title{Computes the second order basis function
for a set of values at a given set of knots.}
\usage{
bs2f(x, x0)
}
\arguments{
\item{x}{set of values}

\item{x0}{set of knots}
}
\description{
Computes the second order basis function
for a set of values at a given set of knots.
}
\examples{
## set of knots
x0 <- 0:5
## set of points to evaluate
x1 <- seq(0, 5, 0.1)
B <- bs2f(x1, x0)
plot(x1, B[,1], type='n', xlab='x', ylab='f(x)')
for (j in 1:ncol(B)) lines(x1, B[,j])
}
