% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_styles.R
\name{snake_case}
\alias{snake_case}
\alias{period.case}
\alias{camelCase}
\title{Convert variable names into one of three styles.}
\usage{
snake_case(df)

period.case(df)

camelCase(df)
}
\arguments{
\item{df}{The dataframe, tibble, etc, with the variables you want to tidy.}
}
\value{
A dataframe, tibble, etc, with snake_case style variable names.

A dataframe, tibble, etc, with period.case style variable names.

A dataframe, tibble, etc, with camelCase style variable names.
}
\description{
Convert variable names to snake_case.

Convert variable names to period.case.

Convert variable names to camelCase.
}
