% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{change_case}
\alias{change_case}
\title{Changing variable cases.}
\usage{
change_case(df, case_style = c("snake_case", "camelCase", "period.case"))
}
\arguments{
\item{df}{The dataframe, tibble, data.table, etc, with the variables you
want to tidy.}

\item{case_style}{The style to convert variable names to. Accepts one of
\code{"snake_case"}, \code{"camelCase"} and \code{"period.case"}.
Defaults to \code{"snake_case"}.}
}
\value{
A dataframe, tibble, data.table, etc, with new and consistent cases
for variable names.
}
\description{
A wrapper for the \code{snake_case()}, \code{camelCase()}, and
\code{period.case()} functions.
}
\examples{
\dontrun{

df <- change_case(df)

}
}
