% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateable.R
\name{dateable}
\alias{dateable}
\title{A function that locates all variables with "date" in the name (ignoring case), and converts them to POSIXct. Works on both dates and datetimes.}
\usage{
dateable(df)
}
\arguments{
\item{df}{The tibble with the dates to be dated.}
}
\description{
A function that locates all variables with "date" in the name (ignoring case), and converts them to POSIXct. Works on both dates and datetimes.
}
