% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emisc-package.R
\docType{package}
\name{emisc}
\alias{emisc}
\alias{emisc-package}
\title{emisc: Evan Odell's various functions}
\description{
emisc: A collection of various R functions I've created that may or may not be useful to anyone else, and that may or may not duplicate existing work.
}
