% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidplot.R
\name{epidplot}
\alias{epidplot}
\title{epidplot}
\usage{
epidplot(
  x,
  y,
  which = 1:3,
  leg.args = list(x = "topleft", legend = c("accumulated", "new", "smoothed new"), col
    = c(1, 2, 2), pch = c(19, 8, NA), lty = c(NA, NA, 1), bty = "n"),
  lxlab = list("", "", ""),
  lylab = list("Cases", "Absolute growth", "Growth velocity"),
  ask = prod(par("mfcol")) < length(which) && dev.interactive(),
  ...
)
}
\arguments{
\item{x}{a date or numeric vector}

\item{y}{the cummulated number of cases}

\item{which}{to specify which plots will be produced.
See details.}

\item{leg.args}{a named list with arguments passed to
the legend of the first plot.}

\item{lxlab}{a list with the xlab for each plot.}

\item{lylab}{a list with the xlab for each plot.}

\item{ask}{logical if ‘TRUE’, the user is _ask_ed
before each plot, see \code{par(ask=.)}.}

\item{...}{additional arguments passed to \code{hist}}
}
\value{
up to three plots.
}
\description{
`epidplot` visualize data
from an epidemic outbreak
}
\details{
The first plot shows the series of accumulated cases,
the number of new cases along with a line which is the
smoothed version of the new cases counts.
The smoothed is computed using the \code{mgcv:::gam}
function with \code{mgcv:::s()} on \code{x} and
\code{family=poisson()}.
The second plot is the, numerical, derivative
of the fitted in the previous plot.
The third plot is the, numerical, derivative
of the fist derivative in the previous plot.
}
