% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussApprox.R
\name{GaussApprox}
\alias{GaussApprox}
\title{GaussApprox
\code{GaussApprox} computes the Gaussian approximation
given the likelihood function, design matrix and
the prior precision matrix.}
\usage{
GaussApprox(f, A, Q, k = 5, h = .Machine$double.eps^0.2, ...)
}
\arguments{
\item{f}{the function that computes the likelihood}

\item{A}{the design matrix}

\item{Q}{the prior precision matrix}

\item{k}{the number of iterations}

\item{h}{the change around the avaluation point}

\item{...}{additional arguments to f}
}
\description{
GaussApprox
\code{GaussApprox} computes the Gaussian approximation
given the likelihood function, design matrix and
the prior precision matrix.
}
\examples{
data(Tokyo, package='INLA')
if (any(ls()=='Tokyo')) {
  llfbt <- function(x)
    dbinom(Tokyo$y, Tokyo$n, plogis(x), log=TRUE)
  n <- nrow(Tokyo)
  library(Matrix)
  Rt <- crossprod(diff(Diagonal(n)))
  tau <- 14500
  ga <- GaussApprox(llfbt, Diagonal(n), Rt*tau)
  with(Tokyo, plot(time, y/n, pch=8))
  lines(ga$mu)
}
}
