% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greatCircleDistance.R
\name{gcDist}
\alias{gcDist}
\title{Great circle distance from longlat coordinates}
\usage{
gcDist(x1, y1, x2, y2, R = 6371)
}
\arguments{
\item{x1}{longitude of the first location set}

\item{y1}{latitude of the second location set}

\item{x2}{longitude of the first location set}

\item{y2}{latitude of the second location set}

\item{R}{the radius. Default is 6371 gives the
the resulting distance in kilometers.}
}
\description{
given the coordinates in the longitude
and latitude it computes the distance between these
points. The default distance computed is in kilometers.
}
\examples{
\dontrun{
## Consider some locations
locs <- rbind(
  Paris=c(2.3295489, 48.8588377),
  Berlin=c(13.4072169, 52.5067614),
  Curitiba=c(-49.2582225, -25.4420471),
  Tokyo=c(139.6708224, 35.6681625))
locs

apply(locs, 1, function(xy)
  gcDist(xy[1], xy[2], locs[,1], locs[,2]))
}
}
