% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htable.R
\name{htable}
\alias{htable}
\title{Frequency table for a numeric vector}
\usage{
htable(x, sep = " to ", digits = 2, ...)
}
\arguments{
\item{x}{numeric vector.}

\item{sep}{character to be used to build the
rownames from the breaks.}

\item{digits}{integer used to format the breaks.}

\item{...}{additional arguments passed to \code{hist}}
}
\value{
a table of frequencies.
}
\description{
Uses the \code{hist} function
to build a frequency table.
}
\examples{
htable(rnorm(400))
}
