% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inla.stats}
\alias{inla.stats}
\title{inla.stats
\code{inla.stats} is to extract dic, waic and cpo
summaries from an output of an object returned
by the 'inla' function of the 'INLA' package.}
\usage{
inla.stats(r, idx)
}
\arguments{
\item{r}{the result of 'inla'}

\item{idx}{integer vector to specify which elements
of the local dic, waic and cpo is to be considered.}
}
\description{
inla.stats
\code{inla.stats} is to extract dic, waic and cpo
summaries from an output of an object returned
by the 'inla' function of the 'INLA' package.
}
