% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlaStats.R
\name{inlaStats}
\alias{inlaStats}
\title{inlaStats
\code{inlaStats} is to extract dic, waic and cpo
summaries from an output of an object returned
by the 'inla' function of the 'INLA' package.}
\usage{
inlaStats(r, idx)
}
\arguments{
\item{r}{the result of 'inla'}

\item{idx}{integer vector to specify which elements
of the local dic, waic and cpo is to be considered.}
}
\description{
inlaStats
\code{inlaStats} is to extract dic, waic and cpo
summaries from an output of an object returned
by the 'inla' function of the 'INLA' package.
}
