% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgsub.R
\name{mgsub}
\alias{mgsub}
\title{Multiple gsub Replacement}
\usage{
mgsub(x, replace_list)
}
\arguments{
\item{x}{a character vector where matches are sought, or an object which can
be coerced by as.character to a character vector. Long vectors are
supported.}

\item{replace_list}{A list of at least length one where each element is of
length 2. Each two element vector has the 'old' pattern and the 'new'
replacement which are passed to gsub via Reduce.}
}
\value{
a character vector of the same length and with the same attributes
as x (after possible coercion to character). Elements of character vectors x
which are not substituted will be returned unchanged. See ?gsub for further
details.
}
\description{
Multiple gsub replacement: Replaces the need for multiple,
consecutive gsub functions. If you have more than once replacement rule 
then this convenience function is appropriate to use.
}
\note{
Code idea taken from:
http://stackoverflow.com/questions/15253954/replace-multiple-arguments-with-gsub
}
\examples{
ex_string = 'Here are the two dates: @startdate, @enddate'
ex_list = list(c('@startdate', '10/1/2010'), c('@enddate', '9/30/2012'))
mgsub(x = ex_string, replace_list = ex_list)

}

