% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysummary.R
\name{mysummary}
\alias{mysummary}
\title{mysummary}
\usage{
mysummary(x, p = c(0, 0.025, 0.25, 0.5, 0.75, 0.975, 1))
}
\arguments{
\item{x}{numeric vector or factor}

\item{p}{percentiles for computing quantiles
when \code{x} is a numeric vector.}
}
\value{
a named vector with the length of \code{x},
number of NA`s and mean, sd an quantiles if \code{x}
is numeric or frequencies if \code{x} is factor.
}
\description{
`mysummary` provides a more complete summary
}
