% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NScases.R
\docType{data}
\name{NScases}
\alias{NScases}
\title{Confirmed cases of COVID19 in
Nova Scotia, Canada.}
\format{
An object of class \code{"list"}
containing 'day' (in Date format) and
'cases', the integer vector of daily counts.
}
\source{
\href{https://novascotia.ca/coronavirus/}{Oficial website}
}
\usage{
data(NScases)
}
\description{
Confirmed cases of COVID19 in
Nova Scotia, Canada.
}
\examples{
data(NScases)
with(NScases, plot(day, cases, pch=8))

}
\keyword{datasets}
