% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSCcases.R
\docType{data}
\name{NSCcases}
\alias{NSCcases}
\title{Confirmed cases of COVID19 in
Nova Scotia, Canada.}
\format{
An object of class \code{"list"}
containing 'day' (in Date format) and
'cases', the integer vector of daily counts.
}
\source{
\href{https://novascotia.ca/coronavirus/}
}
\usage{
data(NSCcases)
}
\description{
Confirmed cases of COVID19 in
Nova Scotia, Canada.
}
\examples{
data(NSCcases)
with(NSCcases, plot(day, cases, pch=8))

}
\keyword{datasets}
