% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar2.R
\name{ar2q}
\alias{ar2q}
\title{ar2q}
\usage{
ar2q(n, a)
}
\arguments{
\item{n}{is the size of the precision matrix}

\item{a}{length three vector with the parameters
a_0, a_1 and a_2 in
a_0 y_t + a_1 y_{t-1} + a_2 y_{t-2} = w_t
stochastic difference equation.}
}
\description{
ar2q
}
\section{Functions}{
\itemize{
\item \code{ar2q}: matrix.
}}

\examples{
ar2q(10, c(1, 1, -0.5))
}
