% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_exists.R
\name{rm_exists}
\alias{rm_exists}
\title{Remove if object exists}
\usage{
rm_exists(vector, envir = parent.frame())
}
\arguments{
\item{vector}{A vector of strings references the objects that you want to have removed.}

\item{envir}{The environment this should be completed within.}
}
\value{
Nothing is returned but the removal of the object.
}
\description{
A function which tests object(s) existence, and then removes it
if it exists.
}
\note{
We will remove rm.exists on or after 11/1/2017; replaced with
rm_exists.
}
\examples{
A <- 1
B <- 2
D <- 4
rm_exists(vector = LETTERS[1:4])
}

