% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_non_alpha.R
\name{strip_non_alpha}
\alias{strip_non_alpha}
\title{strip_non_alpha}
\usage{
strip_non_alpha(df)
}
\arguments{
\item{df}{The dataframe, tibble, etc, with the variables you want to tidy.}
}
\value{
A dataframe, tibble, etc, with new and consistent cases for 
variable names.
}
\description{
Strip out most non-alphanumeric characters from variables. Removes all 
alphanumeric characters except for periods and underscores ('.' and '_'), 
as these are often used in variable names.
}
\examples{
\dontrun{

names(df)
#[1] "name_with_brackets[]"                "$$name_with_dollar_sign"            
#[3] "name_&&with&&_ambersand::_and_colon" "name_'with\\"_quote" 

df <- strip_non_alpha(df)

names(df)
#[1] "name_with_brackets"             "name_with_dollar_sign"    
#[2] "name_with_ambersand_and_colon"  "name_with_quote" 

}
}
