% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_variables.R
\name{tidy_variables}
\alias{tidy_variables}
\title{Strip out most non-alphanumeric characters from variables}
\usage{
tidy_variables(df)
}
\arguments{
\item{df}{The dataframe, tibble, etc, with the variables you want to tidy.}
}
\value{
A dataframe, tibble, etc, with new and consistent cases for variable names.
}
\description{
Removes all alphanumeric characters except for periods and underscores ('.' and '_'), as these are often used in variable names.
}
\examples{
\dontrun{

df <- tidy_variable(df)

}
}
