% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeLines.R
\name{timeLines}
\alias{timeLines}
\title{Function to plot multiple time series.}
\usage{
timeLines(
  w,
  x = 1:ncol(w),
  xlim = range(x),
  ylim = range(w, na.rm = TRUE),
  xlab = "",
  ylab = "",
  axis1args = list(at = pretty(x)),
  axis2args = list(at = pretty(ylim), las = 1),
  col = NULL,
  x.expand = 0.35,
  y.labs = rownames(w),
  y.ylabs = NULL,
  yh = 0.7 * diff(range(w[, ncol(w)], na.rm = TRUE))/nrow(w),
  ...
)
}
\arguments{
\item{w}{numeric matrix, each line a time series data}

\item{x}{vector with the x-axis values}

\item{xlim}{numeric length two vector with x-axis limits}

\item{ylim}{numeric length two vector with y-axis limits}

\item{xlab}{label for the x-axis}

\item{ylab}{label for the y-axis}

\item{axis1args}{arguments list for x-axis}

\item{axis2args}{arguments list for y-axis}

\item{col}{colors for each line drawn}

\item{x.expand}{numeric to be used as fraction of
diff(xlim) as extension of it}

\item{y.labs}{labels to be added like a legend}

\item{y.ylabs}{numeric vector with y-values to
place y.labs}

\item{yh}{numeric used as h in cutree(hclust(...))
to arrange the y.labs towards a more
equally spaced sequence over y-limits.
NOTE: This is only used if y.ylabs is not NULL.}

\item{...}{further arguments passed further}
}
\value{
plot
}
\description{
Function to plot multiple time series.
}
\examples{
timeLines(t(Seatbelts[,3:4]))
}
