% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfR.R
\name{transfR}
\alias{transfR}
\title{Function for Real to Real transform
of Real to Positive functions}
\usage{
transfR(x, transf)
}
\arguments{
\item{x}{numeric}

\item{transf}{character name of the function
that implements the transformation to be applied.}
}
\value{
numeric with the transformed input
}
\description{
Compute the the desired transformation
for the abs of the input and multiply
the output by the sign of the input.
}
\examples{
transfR(-5:5, 'sqrt')
transfR(-5:5, 'log10')
plot(function(x) transfR(x, 'log'), -10, 10)
plot(function(x) transfR(x, 'log2'), -10, 10,
     col=2, add=TRUE)
plot(function(x) transfR(x, 'log10'), -10, 10,
     col=4, add=TRUE)
}
