% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stri_replace_vector.R
\name{vector_stri_replace_all_fixed}
\alias{vector_stri_replace_all_fixed}
\alias{vector_stri_replace_all_fixed_}
\alias{vector_stri_replace_all_regex}
\alias{vector_stri_replace_all_regex_}
\title{vector_stri_replace}
\usage{
vector_stri_replace_all_fixed(x, ..., progress = TRUE)

vector_stri_replace_all_fixed_(x, from, to, progress = TRUE)

vector_stri_replace_all_regex(x, ..., progress = TRUE)

vector_stri_replace_all_regex_(x, from, to, progress = TRUE)
}
\arguments{
\item{x}{Character vector with strings to search in}

\item{...}{One or more strings for search patterns and their 
replacement pattern.}

\item{progress}{If \code{TRUE}, displays a progress bar. Defaults to TRUE.}

\item{from}{A character vector defining search patterns}

\item{to}{A character vector with replacements for matched patterns}
}
\value{
A character vector with strings replaced.
}
\description{
A simple wrapper around \code{replace_all_fixed_vector}
}
\details{
When using the \code{vector_stri_replace_all_fixed_}, which is suitable 
for programming, the variables \code{from} and \code{to} must be character 
vectors of equal length.
}
