% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x2rgb.R
\name{x2rgb}
\alias{x2rgb}
\title{x2rgb}
\usage{
x2rgb(x, breaks = NULL, u = FALSE, reverse = FALSE, ...)
}
\arguments{
\item{x}{numeric vector}

\item{breaks}{non decreasing numeri vector}

\item{u}{logical. Uses the rank of \code{x} as \code{x}.}

\item{reverse}{logic. If FALSE colors are blue for lower
values of \code{x} and red for hight values of \code{x}.
If TRUE, colors are red for lower values of \code{x} and
blue for blue for hight values of \code{x}.}

\item{...}{additional parameters passed to \code{\link{rgb}}.}
}
\value{
'transparent' if NA or the output of the
\code{\link{rgb}} function
}
\description{
`x2rgb` create rgb colors from a numeric vector
}
\section{Warning}{

'transparent' is returned for NA
}

\examples{
plot(0:5, pch=19, col=x2rgb(0:5))
plot(0:5, pch=19, col=x2rgb(0:5, c(0,3,6)))
}
\seealso{
\code{\link{rgb}}
}
