% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar2.R
\name{ar2acf}
\alias{ar2acf}
\title{Functions for the ar2 model}
\usage{
ar2acf(a1, a2, k)
}
\arguments{
\item{a1}{first difference parameter.}

\item{a2}{second difference parameter or
partial correlation of order two.}

\item{k}{max lag to compute the acf.}
}
\value{
vector of length k or matrix of k columns
with autocorrelation up to lag k for each
parameter pairs, each line if matrix.
}
\description{
Compute the autocorrelation function
given the partial correlation coefficients.
}
\section{Functions}{
\itemize{
\item \code{ar2acf()}: marginal correlation.

}}
\examples{
ar2acf(1, -0.5, 10)
ar2acf(c(1,1.8), c(-0.5, -0.5), 10)
}
