% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emisc-package.R
\docType{package}
\name{emisc}
\alias{emisc}
\alias{emisc-package}
\title{emisc: Evan Odell's various functions}
\description{
emisc: A collection of various R functions I've created. They may or may 
not be useful to anyone else, and may or may not  duplicate existing work. 
Includes functions to remove byte order marks  (BOMs), write data to LaTeX 
as lines rather than tables, and changing variable naming conventions.
}
