% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidplot.R
\name{epidplot}
\alias{epidplot}
\title{Visualize an epdemic outbreak}
\usage{
epidplot(
  x,
  y,
  log10 = TRUE,
  which = list(1:2, c(3, 5), 4, 6),
  w = exp(-sqrt(1:14)) * 0.686,
  leg.args = NULL,
  lxlab = NULL,
  lylab = NULL,
  ask = prod(par("mfcol")) < length(which) && dev.interactive(),
  ...
)
}
\arguments{
\item{x}{a date or numeric vector}

\item{y}{the cummulated number of cases}

\item{log10}{length one or two logical to
indicate if accumulated and new series are
shown in the log10 transformation.}

\item{which}{integer vector or list of integers
to specify which plots will be produced in each
figure. See detais.}

\item{w}{numeric vector to compute the expected
number of new cases given past cases. See details.}

\item{leg.args}{a named list with arguments
passed to the legend of the first plot.}

\item{lxlab}{a list with the xlab for each plot.}

\item{lylab}{a list with the xlab for each plot.}

\item{ask}{logical if ‘TRUE’, the user is _ask_ed
before each plot, see \code{par(ask=.)}.}

\item{...}{additional arguments passed to \code{hist}}
}
\value{
up to four plots.
}
\description{
visualize the cummulated time
series and its derivative (news).
Add smoothed lines of the news.
Additional plots shows the 1st and 2nd
derivatives of the smoothed, interpreted
as the absolute increase and velocity.
}
\details{
The first plot shows the series of accumulated cases,
the number of new cases along with a line which is the
smoothed version of the new cases counts.
The smoothed is computed using the \code{mgcv:::gam}
function with \code{mgcv:::s()} on \code{x} and
\code{family=poisson()}.
The second plot is the, numerical, derivative
of the fitted in the previous plot.
The third plot is the, numerical, derivative
of the fist derivative in the previous plot.
The fourth plot is a smoothed version of R_t.
This is computed as R_t = s_t/E_t, where
s_t is smoothed version of the (new) series,
E_t = sum_j w_j s_{t-j}.
}
\examples{
## COVID19 cases in Nova Scotia, Canada
ns.cases <- c(5, 7, 12, 14, 15, 21, 28, 41,
 51, 68, 73, 90, 110, 122, 127, 147, 173,
 193, 207, 236, 262, 293, 310, 310, 342,
 407, 428, 445, 474, 517, 549, 579, 606,
 649, 675, 721, 737, 772, 827, 850, 865,
 873, 900, 915, 935, 947, 959, 963, 971,
 985, 991, 998, 1007, 1008, 1011, 1018,
 1019, 1020, 1024, 1026, 1034, 1037, 1040)
date <- as.Date('2020-03-15') + 1:length(ns.cases)
pw <- pgamma(0:14, shape=(5/3)^2, scale=3^2/5)
w <- diff(pw)/sum(diff(pw))
par(mfrow=c(2,2), mar=c(2,3,1,1),
    mgp=c(2, 0.5, 0), las=1)
epidplot(date, ns.cases, logbase=10, w=w)
}
