% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidplot.R
\name{epidplot}
\alias{epidplot}
\title{Visualize an epdemic outbreak}
\usage{
epidplot(
  x,
  y,
  which = 1:4,
  w,
  leg.args = list(x = "topleft", legend = c("accumulated", "new", "smoothed new"), col
    = c(1, 2, 2), pch = c(19, 8, NA), lty = c(NA, NA, 1), bty = "n"),
  lxlab = list("", "", "", ""),
  lylab = list("Cases", "Absolute growth", "Growth velocity", "Reproduction number"),
  ask = prod(par("mfcol")) < length(which) && dev.interactive(),
  ...
)
}
\arguments{
\item{x}{a date or numeric vector}

\item{y}{the cummulated number of cases}

\item{which}{integer vector to pecify which
plots will be produced. See details.}

\item{w}{numeric vector to compute the expected
number of new cases given past cases. See details.}

\item{leg.args}{a named list with arguments
passed to the legend of the first plot.}

\item{lxlab}{a list with the xlab for each plot.}

\item{lylab}{a list with the xlab for each plot.}

\item{ask}{logical if ‘TRUE’, the user is _ask_ed
before each plot, see \code{par(ask=.)}.}

\item{...}{additional arguments passed to \code{hist}}
}
\value{
up to four plots.
}
\description{
visualize the cummulated time
series and its derivative (news).
Add smoothed lines of the news.
Additional plots shows the 1st and 2nd
derivatives of the smoothed, interpreted
as the absolute increase and velocity.
}
\details{
The first plot shows the series of accumulated cases,
the number of new cases along with a line which is the
smoothed version of the new cases counts.
The smoothed is computed using the \code{mgcv:::gam}
function with \code{mgcv:::s()} on \code{x} and
\code{family=poisson()}.
The second plot is the, numerical, derivative
of the fitted in the previous plot.
The third plot is the, numerical, derivative
of the fist derivative in the previous plot.
The fourth plot is a smoothed version of R_t.
This is computed as R_t = s_t/E_t, where
s_t is smoothed version of the (new) series,
E_t = sum_j w_j s_{t-j}.
}
\examples{
ns.cases <- c(5, 7, 12, 14, 15, 21, 28, 41,
 51, 68, 73, 90, 110, 122, 127, 147, 173,
 193, 207, 236, 262, 293, 310, 310, 342,
 407, 428, 445, 474, 517, 549, 579, 606,
 649, 675, 721, 737, 772, 827, 850, 865,
 873, 900, 915, 935, 947, 959, 963, 971,
 985, 991, 998, 1007, 1008, 1011, 1018, 1019)
date <- as.Date('2020-03-15') + 1:length(ns.cases)
w <- pgamma(0:14, shape=(7/4)^2, scale=4^2/7)
par(mfrow=c(2,2), mar=c(2,3,1,1),
    mgp=c(2, 0.5, 0), las=1)
epidplot(date, ns.cases, w=w)
}
