% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htable.R
\name{htable}
\alias{htable}
\title{\code{htable} provides a frequency table using
the \code{hist} function internally.}
\usage{
htable(x, ...)
}
\arguments{
\item{x}{numeric vector}

\item{...}{additional arguments passed to \code{hist}}
}
\value{
a table of frequencies.
}
\description{
\code{htable} provides a frequency table using
the \code{hist} function internally.
}
\examples{
htable(rnorm(100))
}
