% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_markup.R
\name{latex_markup}
\alias{latex_markup}
\title{latex_markup}
\usage{
latex_markup(df, ...)
}
\arguments{
\item{df}{The dataframe, data.table, tibble, etc, containing the data to add LaTeX markup to.}

\item{...}{One or more unquoted, comma separated, variable names, with quoted Latex environment names (no backslashes or curly braces)/
#@param markup_style}
}
\description{
Add LaTeX markup to R data.
}
\examples{
\dontrun{

x <- latex_markup(iris, Species="section", Petal.Width="subsection*")#, markup_style="section")

 # >  head(x)
 #    Sepal.Length Sepal.Width Petal.Length       Petal.Width           Species
 #  1          5.1         3.5          1.4 \\\\subsection*{0.2} \\\\section{setosa}
 #  2          4.9         3.0          1.4 \\\\subsection*{0.2} \\\\section{setosa}
 #  3          4.7         3.2          1.3 \\\\subsection*{0.2} \\\\section{setosa}
 #  4          4.6         3.1          1.5 \\\\subsection*{0.2} \\\\section{setosa}
 #  5          5.0         3.6          1.4 \\\\subsection*{0.2} \\\\section{setosa}
 #  6          5.4         3.9          1.7 \\\\subsection*{0.4} \\\\section{setosa}

 
}
}
