% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_functions.R
\name{set_functions}
\alias{set_functions}
\alias{setf}
\title{set functions for multiple column alterations using data.tables}
\usage{
setf(x, j, value, ...)
}
\arguments{
\item{x}{A data.table. Or, set() accepts data.frame, too.

\code{i} Is left out of this function.}

\item{j}{Column name(s) (character) or number(s) (integer) to be assigned
value when column(s) already exist, and only column name(s) if they are to
be added newly. Additionally, assigning \code{j = 'all'} will automatically
reference all columns in the data.table.}

\item{value}{A list of replacement values to assign by reference to
\code{x[i, j]}. In addition, functions can be passed into the value
argument as well with subsquent arguements in the same way as apply
functions.}

\item{...}{Is used to pass arguments into the value "function" if you choose
to apply a function instead of updating the data with a replacement vector
or list. If you choose to place a \code{.} at any point within the
sequence of additional arguements passed into the \code{value} arguement,
the \code{.} will act as a placeholder for the jth column.}
}
\description{
This function updates by reference multiple columns by applying
  a desired function such as \code{numeric}, \code{factor} using functions in
  the same manior as \code{apply()}, or with \code{.} as a placeholder.
}
\section{Notes}{
 These functions are for convenience only.
}

\section{TODO}{

\itemize{
   \item Consider to.function joining.
}
}
\examples{
dt <- d.t(x = 0:10, y = 10:20)
setf(dt, j = c("x","y"), value = paste0,  ., "hello", "would")
dt <- d.t(x = 1:10, y = 11:20)
print(dt)
}

