% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stplot.R
\name{stplot}
\alias{stplot}
\title{Time series plot for each location}
\usage{
stplot(
  x,
  sp,
  d,
  col,
  ce,
  tsub = 3,
  ex = 1,
  ey = 1,
  leg.args = list(x = "bottomleft", lty = 1),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{a matrix with each column containing data
at each time for a set of geographical units}

\item{sp}{an object as defined in the 'sp' package}

\item{d}{a vector to specify a \code{GridTopology}
object to do some subset over the \code{sp} object.}

\item{col}{a vector of colors for each time series.
If not provided, the collors will be defined considering
the relative growth of each time series and series with
similar growth will have similar colors. By default
red color is for the series with most positive growth
and blue for those with least growth.}

\item{ce}{vector of coordinates to center the time
the time series. If not provided, the \code{coordinates}
output applied to \code{sp}.}

\item{tsub}{the number of temporal subdivisions
to evaluate the growth of each time series.}

\item{ex}{is a relative expansion factor for
the x-axis extent}

\item{ey}{is a relative expansion factor for
the y-axis extent}

\item{leg.args}{legend arguments}

\item{verbose}{logical to specify if some steps
are being reported during the process work}

\item{...}{additional arguments passed to
\code{line} used to plot each time series.}
}
\description{
given a matrix where each line is a
time series of each spatial unit, the time series
are shown over the map
}
\examples{
data(measlesWeserEms, package='surveillance')
if (any(ls()=='measlesWeserEms'))
  stplot(t(measlesWeserEms@observed),
         measlesWeserEms@map)
}
