% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tidy_bom}
\alias{tidy_bom}
\title{Internal function for \code{\link{strip_bom}}}
\usage{
tidy_bom(df)
}
\arguments{
\item{df}{The API response to remove BOMs from.}
}
\description{
Internal function for \code{\link{strip_bom}}
}
