% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGMRF.R
\name{updateGMRF}
\alias{updateGMRF}
\title{Computes the posterior Gaussian distribution
given the (mixed) design and precision matrices.}
\usage{
updateGMRF(y, Qe, A, Qx)
}
\arguments{
\item{y}{the outcome}

\item{Qe}{the error precision matrix}

\item{A}{the design matrix}

\item{Q}{the prior precision matrix for the
latent random (Markov) random field}
}
\description{
Computes the posterior Gaussian distribution
given the (mixed) design and precision matrices.
}
\examples{
## Consider the Orange dataset
## intercept plus linear trend over age with
## one tred for each tree
##   y_{ij} = a + b_j age_i + error_{ij}
y <- Orange$circumference
z0 <- model.matrix(~Tree-1, Orange)
A <- cbind(1, Orange$age, z0*Orange$age)
Qx <- Diagonal(2 + 5, c(0, 0, rep(3000, 5)))
Qe <- Diagonal(nrow(Orange), 0.01)
result <- updateGMRF(y, Qe, A, Qx)
print(cbind(x=result$mu,
   sd=sqrt(diag(chol2inv(result$L)))))

## compare with INLA result
\dontrun{
  ff <- circumference ~ f(Tree, age, model='iid')
  res <- inla(ff, data=Orange)
  res$summary.fixed[, 1:2]
  res$summary.random$Tree[, 1:3]
}
}
